requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.19.0
alembic==1.13.1
icmplib==3.0.4
apscheduler==3.10.4
httpx==0.26.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
jinja2==3.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
itsdangerous==2.1.2
certifi==2025.1.31
